backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "ffmpeg-python>=0.2.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
